requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-webpack-loader==3.0.1
djangorestframework==3.14.0
pandas==2.2.1
sqlparse==0.4.4
tqdm==4.66.2
camelot-py>=0.11.0
opencv-python==4.9.0.80
ghostscript